--- src/app/property-management/property-management.component.html
@@ -1,28 +1,0 @@
-<div id="container" class="d-flex flex-column gap-3 justify-content-center align-items-center">
-  <h3 id="page-title">Property Management</h3>
-  <section class="w-75">
-    <button class="btn btn-primary w-100" id="button-add-property" type="button"
-      data-target="#exampleModal" (click)="openModal()"><i class="bi bi-plus"></i> Add Property</button>
-    <div class="accordion accordion-flush" id="accordionFlushExample">
-      <div class="accordion-item" *ngFor="let propertyHeader of propertiesHeaderList">
-        <h2 class="accordion-header">
-          <button class="accordion-button collapsed" type="button" (click)="getPropertyBody(propertyHeader.id)" data-bs-toggle="collapse" [attr.data-bs-target]="'#flush-collapseThree' + propertyHeader.id" aria-expanded="false" aria-controls="flush-collapseThree">
-            {{propertyHeader.typologyId}} - {{propertyHeader.cityId}} I {{propertyHeader.title}}
-          </button>
-        </h2>
-        <div [id]="'flush-collapseThree' + propertyHeader.id" class="accordion-collapse collapse" data-bs-parent="#accordionFlushExample">
-          <div class="accordion-body">
-            <p><strong>Address:</strong> {{propertyBody?.address}}</p>
-            <p><strong>Postal Code:</strong> {{propertyBody?.zipCode}}</p>
-            <p><strong>Construction Year:</strong> {{propertyBody?.build_Date}}</p>
-            <p><strong>Square Meters:</strong> {{propertyBody?.squareMeter}}</p>
-            <p><strong>Energetic Class:</strong> {{propertyBody?.energyClass}}</p>
-            <button class="btn btn-danger" type="button" (click)="deleteModal(propertyHeader.id)">
-              <i class="bi bi-trash"></i>
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  </section>
-</div>
