+++ src/app/real-estate-management/real-estate-management.component.html
@@ -17,7 +17,9 @@
             <p><strong>Build Date:</strong> {{realEstateBody?.build_Date}}</p>
             <p><strong>Square Meters:</strong> {{realEstateBody?.squareMeter}}</p>
             <p><strong>Energy Class:</strong> {{realEstateBody?.energyClass}}</p>
-            <button class="btn btn-danger" type="button"><i class="bi bi-trash"></i></button>
+            <button class="btn btn-danger" type="button" (click)="deleteModal(realEstateHeader.id)">
+              <i class="bi bi-trash"></i>
+            </button>
           </div>
         </div>
       </div>
